+++ backend/services/status.ts
@@ -1,9 +1,19 @@
 import { getCounterTable } from "../models/status";
+import { getVersion } from "../util/version";
 
-export function getCounter(table: string): Promise<Number> {
+export function getCurrentStatus(): Promise<any> {
+    let tables = ["user", "recipe", "tag"];
     return new Promise((resolve, reject) => {
-        getCounterTable(table).then(counter => {
-            resolve(counter);
+        Promise.all([getCounterTable(tables[0]), getCounterTable(tables[1]), getCounterTable(tables[2])]).then(counter => {
+            let status = {
+                "version": getVersion(),
+                "userCount": counter[0],
+                "recipeCount": counter[1],
+                "commentsCount": "-",
+                "tagsCount": counter[2],
+                "categoriesCount": "-"
+            }
+            resolve(status);
         }).catch(err => {
             reject(err);
         });
